+++ include/Hazmat/Math/Matrix/Identity.h
@@ -29,7 +29,7 @@ BEGIN_HAZMAT_NAMESPACE
 namespace Priv
 {
 
-template <size_t DIM, typename T>
+template <int DIM, typename T>
 class Identity;
 
 template <typename T>
@@ -93,7 +93,7 @@ public:
 
 }
 
-template <size_t DIM, typename T>
+template <int DIM, typename T>
 inline
 Matrix<DIM, T>& Identity(Matrix<DIM, T>& a)
 {
