+++ include/Hazmat/Math/Vector/Sub.h
@@ -29,7 +29,7 @@ BEGIN_HAZMAT_NAMESPACE
 namespace Priv
 {
 
-template <size_t DIM, typename T>
+template <int DIM, typename T>
 class Subtraction;
 
 template <typename T>
@@ -79,7 +79,7 @@ public:
 
 }
 
-template <size_t DIM, typename T>
+template <int DIM, typename T>
 inline
 Vector<DIM, T>& Sub(const Vector<DIM, T>& u,
                     const Vector<DIM, T>& v,
@@ -88,7 +88,7 @@ Vector<DIM, T>& Sub(const Vector<DIM, T>& u,
     return Priv::Subtraction<DIM, T>::compute(u, v, w);
 }
 
-template <size_t DIM, typename T>
+template <int DIM, typename T>
 inline
 Vector<DIM, T>& operator-=(      Vector<DIM, T>& u,
                            const Vector<DIM, T>& v)
@@ -96,7 +96,7 @@ Vector<DIM, T>& operator-=(      Vector<DIM, T>& u,
     return Sub(u, v, u);
 }
 
-template <size_t DIM, typename T>
+template <int DIM, typename T>
 inline
 Vector<DIM, T> operator-(      Vector<DIM, T>  u,
                          const Vector<DIM, T>& v)
